Add GitHub Action Step Name (#9)
# Pull Request

## Description

This pull request includes a minor update to the
`.github/workflows/sync-labels.yml` file to improve the clarity of a
workflow step by adding a descriptive `name` field.

*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L25-R26):
Added a `name` field ("Sync labels") to the step using
`micnncim/action-label-syncer@v1.3.0` for better readability and
identification in the GitHub Actions workflow.