Swap SuperLinter to Full Version (#27)
# Pull Request

## Description

This pull request includes updates to configuration files to ensure
compatibility and improve maintainability. The most important changes
involve updating dependencies and configuration versions.

### Dependency and Configuration Updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L27-R28):
Updated the `uses` directive for the Super-Linter action to remove the
`/slim` suffix, aligning with the latest recommended usage.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` of Lefthook from `1.11.2` to `1.11.13` to
ensure compatibility with newer features and bug fixes.