Migrate Sync Labels to Common Sync Labels (#35)
# Pull Request

## Description

This pull request updates the `.github/workflows/sync-labels.yml` file
to simplify the workflow configuration by replacing custom steps with a
reusable workflow and adjusting permissions.

### Workflow simplification:
* Replaced the custom steps for syncing labels with a reusable workflow
from `JackPlowman/reusable-workflows` to streamline the configuration.

### Permissions adjustments:
* Updated `permissions` to `{}` at the top level and set specific
permissions (`contents: read`, `pull-requests: write`) for the
`configure-labels` job to align with the principle of least privilege.

### Trigger updates:
* Added `workflow_dispatch` as a trigger to allow manual execution of
the workflow.